Allow other plugins to provide current filename
When displaying the Open File dialog, the name of the currently displayed
file is used to initialize the dialog. This commit asks the currently
focussed plugin first to provide the current filename, and if that fails,
the current filename of the Editor is used.

The `sig_focused_plugin_changed` signal in the previous commit is used
to find which plugin currently has focus.