Variable explorer: Show error if missing package in console env
If retrieving a variable from the kernel fails due to a
ModuleNotFoundError raised by call_kernel(), then show an error
message instructing the user to install the missing module in the
console environment.

This error occurs if a user tries to retrieve a Polars dataframe,
because the kernel will transform it to a Pandas dataframe and
this fails if pandas is not installed in the kernel environment.